+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 * Pre-hotreload `cartridge` support (older than 2.4.0) (PR #341).
+* Tarantool version-dependent features for 3.x (PR #344).
 
 ## [1.0.0] - 02-02-23